requirements.txt
httplib2
Django==1.3
django-dajaxice

